Fixed the migrations for creating vehicles and customers tables.

+++ database/migrations/2019_12_11_074003_create_models_table.php
@@ -16,7 +16,7 @@ class CreateModelsTable extends Migration
         Schema::create('models', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name');
-            $table->unsignedInteger('brand_id');
+            $table->unsignedBigInteger('brand_id');
             $table->timestamps();
 
             $table->foreign('brand_id')->references('id')->on('brands');
